internal/labeler: Sync the do-not-merge/kind-invalid label correctly
The labeler was previously not removing the "do-not-merge/kind-invalid"
label if a PR was updated from an invalid/missing kind state to a
valid kind state.

This commit updates the `verifyKinds` function to explicitly mark
"do-not-merge/kind-invalid" for removal if it exists on a PR and
the kinds specified in the PR body are determined to be valid.

Signed-off-by: timflannagan <timflannagan@gmail.com>